fix(terminal): settle the mirror on the frame's decision, not on a bare false
`shouldApplyWebSessionTabsSnapshot` returned false for two unrelated reasons,
and every settle site read that boolean as staleness. One of them — the
floating workspace, a local synthetic terminal the mirror never writes — has
no accepted store view behind it, so settling on it drains parked resume
waiters against state nobody wrote and lets the sweep launch a replacement
`codex resume` against a session the host still holds.

The predicate now returns a decision that carries `settlesHostMirror` beside
`apply`, minted at each branch, so a future rejection branch cannot inherit a
settle by defaulting. Four sites derive their verdict from it: both hand-built
patchless receipts and both inventory `settles` lists, which named the floating
worktree the patch had filtered out. Inventory counts exclude unmirrored
frames too, so a host with a floating terminal keeps its environment-wide
verdict.

Splits the frame-ordering suite around the shared harness it outgrew.
